+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoRuntimeSettings.java
@@ -675,6 +675,8 @@ public final class GeckoRuntimeSettings extends RuntimeSettings {
       new PrefWithoutDefault<Boolean>("fission.disableSessionHistoryInParent");
   /* package */ final Pref<Boolean> mFetchPriorityEnabled =
       new Pref<Boolean>("network.fetchpriority.enabled", false);
+  /* package */ final Pref<Boolean> mParallelMarkingEnabled =
+      new Pref<Boolean>("javascript.options.mem.gc_parallel_marking", false);
   /* package */ final Pref<Boolean> mCookieBehaviorOptInPartitioning =
       new Pref<Boolean>("network.cookie.cookieBehavior.optInPartitioning", false);
   /* package */ final Pref<Boolean> mCookieBehaviorOptInPartitioningPBM =
@@ -1024,6 +1026,27 @@ public final class GeckoRuntimeSettings extends RuntimeSettings {
    */
   public boolean getFetchPriorityEnabled() {
     return mFetchPriorityEnabled.get();
+  }
+
+  /**
+   * Set the pref to control whether javascript.options.mem.gc_parallel_marking is enabled.
+   *
+   * @param enabled Whether to enable the JS GC Parallel Marking feature. This feature is purely a
+   *     performance feature and should have no noticeable behavior change for the user.
+   * @return This GeckoRuntimeSettings instance
+   */
+  public @NonNull GeckoRuntimeSettings setParallelMarkingEnabled(final boolean enabled) {
+    mParallelMarkingEnabled.commit(enabled);
+    return this;
+  }
+
+  /**
+   * Get whether javascript.options.mem.gc_parallel_marking is enabled.
+   *
+   * @return Whether Parallel Marking is enabled
+   */
+  public boolean getParallelMarkingEnabled() {
+    return mParallelMarkingEnabled.get();
   }
 
   /**